move NaturalIdLogging to package with its only client
